moved TestObsHandler to proper folder.
